requirements.txt
web3==6.15.1
python-dotenv==1.0.1
aiohttp==3.9.3
eth-abi==5.0.1
eth-utils==4.0.0
eth-typing==4.0.0
requests==2.31.0
loguru==0.7.2 